created get-line-at-offset helper and refactoring
