+++ scripts/helpers/get-line-at-offset.js
@@ -1,0 +1,34 @@
+/**
+ * @typedef {Object} Line
+ * @property {string} value
+ * @property {number} index
+ * @property {number} start
+ * @property {number} end
+ */
+
+/**
+ * @param {Array<Line>} lines
+ * @param {number} offset
+ * @return {Line}
+ */
+function getLineAtOffset(lines, offset) {
+  let lower = 0;
+  let upper = lines.length - 2;
+
+  while (lower < upper) {
+    const middle = (lower + (upper - lower) / 2) | 0;
+
+    if (offset < lines[middle].start) {
+      upper = middle - 1;
+    } else if (offset >= lines[middle + 1].start) {
+      lower = middle + 1;
+    } else {
+      lower = middle;
+      break;
+    }
+  }
+
+  return lines[lower];
+}
+
+exports.getLineAtOffset = getLineAtOffset;
